RF: Move addurls() out of plugins
Towards a fix for #4519
